refactor: name change to user.entity.ts
